Add files via upload
Use any Infrared remote control (Receiver or TV remote should work) as a controller for a servo motor by directing the IR beam to the IR receiver (1838).
You are required to make the motor rotate according to the number you send. For example: in order to make the motor rotate at 180 degrees, you send ‘1’ then ‘8’ then ‘0’ then ‘Menu’ or any other function key and the motor should perform the correct rotation. The Volume up and down keys should make the motor angle increase or decrease by 10 degrees. Use an external supply (9V battery) for the motor and attach one of the gears to the motor to show its current position.
NOTE: In order to read the value of each remote control button you have to use an external library:
https://github.com/z3t0/Arduino-IRremote
Please read the library documentation and how to use it since there will be a collision with the RobotIRremote library that comes already with the Arduino software. One approach is to delete the RobotIRremote, the other is to simply rename the downloaded library files and includes another name.
The video should clearly show the pressing of the remote buttons and their effect on the serial monitor.